Struct pallet_bitcoin_vaults::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn do_remove_xpub(who: T::AccountId) -> DispatchResult

Use with caution

pub fn do_insert_vault(vault: Vault<T>) -> DispatchResult

pub fn do_remove_vault( owner: T::AccountId, vault_id: [u8; 32] ) -> DispatchResult

pub fn do_remove_proposal(proposal_id: [u8; 32]) -> DispatchResult

pub fn do_propose(proposal: Proposal<T>) -> DispatchResult

pub fn do_save_psbt( signer: T::AccountId, proposal_id: [u8; 32], signature_payload: BoundedVec<u8, T::PSBTMaxLen> ) -> DispatchResult

pub fn do_finalize_psbt( signer: T::AccountId, proposal_id: [u8; 32], broadcast: bool ) -> DispatchResult

pub fn do_create_proof( signer: T::AccountId, vault_id: [u8; 32], message: Description<T>, psbt: PSBT<T> ) -> DispatchResult

pub fn do_save_proof_psbt( signer: T::AccountId, vault_id: [u8; 32], psbt: PSBT<T> ) -> DispatchResult

pub fn do_finalize_proof( signer: T::AccountId, vault_id: [u8; 32], final_proof: PSBT<T> ) -> DispatchResult

pub fn do_remove_proof(vault_id: [u8; 32])

pub fn get_xpub_status(who: T::AccountId, xpub_hash: [u8; 32]) -> XpubStatus

pub fn vault_comprobations( vault_id: [u8; 32], signer: &T::AccountId ) -> DispatchResult

pub fn do_insert_descriptors( vault_id: [u8; 32], descriptors: Descriptors<T::OutputDescriptorMaxLen>, status: BDKStatus<T::VaultDescriptionMaxLen> ) -> DispatchResult

pub fn do_insert_psbt( proposal_id: [u8; 32], psbt: BoundedVec<u8, T::PSBTMaxLen>, status: BDKStatus<T::VaultDescriptionMaxLen> ) -> DispatchResult

pub fn do_insert_tx_id( proposal_id: [u8; 32], tx_id: Option<BoundedVec<u8, T::VaultDescriptionMaxLen>>, status: BDKStatus<T::VaultDescriptionMaxLen> ) -> DispatchResult

pub fn get_pending_vaults() -> Vec<[u8; 32]>

pub fn get_pending_proposals() -> Vec<[u8; 32]>

pub fn get_accounts_xpubs( accounts: Vec<T::AccountId> ) -> Result<Vec<Vec<u8>>, DispatchError>

pub fn get_proposals_to_finalize() -> Vec<[u8; 32]>

pub fn extract_json_str_by_name( tuple: Vec<(Vec<char>, JsonValue)>, s: &str ) -> Option<Vec<u8>>

pub fn bdk_gen_vault( vault_id: [u8; 32] ) -> Result<(Vec<u8>, Vec<u8>), OffchainStatus>

pub fn gen_vaults_payload_by_bulk( pending_vaults: Vec<[u8; 32]> ) -> Vec<SingleVaultPayload>

pub fn gen_proposal_json_body( proposal_id: [u8; 32] ) -> Result<Vec<u8>, OffchainStatus>

pub fn bdk_gen_proposal( proposal_id: [u8; 32], api_endpoint: Vec<u8>, json_builder: &dyn Fn([u8; 32]) -> Result<Vec<u8>, OffchainStatus> ) -> Result<Vec<u8>, OffchainStatus>

pub fn gen_proposals_payload_by_bulk( pending_proposals: Vec<[u8; 32]>, api_endpoint: Vec<u8>, json_builder: &dyn Fn([u8; 32]) -> Result<Vec<u8>, OffchainStatus> ) -> Vec<SingleProposalPayload>

pub fn gen_finalize_json_body( proposal_id: [u8; 32] ) -> Result<Vec<u8>, OffchainStatus>

pub fn send_ocw_insert_descriptors( generated_vaults: Vec<SingleVaultPayload>, signer: &Signer<T, T::AuthorityId> )

pub fn send_ocw_insert_psbts( generated_proposals: Vec<SingleProposalPayload>, signer: &Signer<T, T::AuthorityId> )

pub fn send_ocw_finalize_psbts( generated_proposals: Vec<SingleProposalPayload>, signer: &Signer<T, T::AuthorityId> )

pub fn chars_to_bytes(v: Vec<char>) -> Vec<u8>

impl<T: Config> Pallet<T>

pub fn set_xpub( origin: OriginFor<T>, xpub: BoundedVec<u8, T::XPubLen> ) -> DispatchResult

XPub insertion

This extrinsic inserts a user-defined xpub as well as in the pallet storage.

Parameters:

xpub: Extended public key, it can be sent with or without fingerprint/derivation path

Considerations

The origin must be Signed and the sender must have sufficient funds free for the transaction fee.

This extrinsic cannot handle a xpub update (yet). if it needs to be updated, remove it first and insert a new one.

pub fn remove_xpub(origin: OriginFor<T>) -> DispatchResult

Xpub removal

Removes the linked xpub from the account which signs the transaction. The xpub will be removed from both the pallet storage and identity registration.

This tx does not takes any parameters.

pub fn create_vault( origin: OriginFor<T>, threshold: u32, description: BoundedVec<u8, T::VaultDescriptionMaxLen>, include_owner_as_cosigner: bool, cosigners: BoundedVec<T::AccountId, T::MaxCosignersPerVault> ) -> DispatchResult

Vault insertion

Inserts the vault on chain. Meant to be used by an offchain worker.

Parameters:

threshold: The number of signatures needed for a proposal to be approved/finalized

description: A small definition. What will the vault be used for?

include_owner_as_cosigner: Add automatically the owner as cosigner

cosigners: The other accounts that will participate in vault proposals.

Considerations

Do not include the vault owner on the cosigners list.

pub fn remove_vault(origin: OriginFor<T>, vault_id: [u8; 32]) -> DispatchResult

Vault removal

Tries to remove vault and all its proposals, only the owner can call this extrinsic.

Parameters:

vault_id: the vault to be removed with all its proposals

Considerations:

Only the vault owner can perform this extrinsic

pub fn propose( origin: OriginFor<T>, vault_id: [u8; 32], recipient_address: BoundedVec<u8, T::XPubLen>, amount_in_sats: u64, description: BoundedVec<u8, T::VaultDescriptionMaxLen> ) -> DispatchResult

Vault transaction proposal

Inserts a proposal on the specified vault.

Parameters:

vault_id: the vault identifier in which the proposal will be inserted

recipient_address: Mainnet address to which the funds will be send

amount_in_sats: Amount to send in satoshis.

description: The reason for the proposal, why do you are proposing this?.

Considerations

Please ensure the recipient address is a valid mainnet address.

pub fn remove_proposal( origin: OriginFor<T>, proposal_id: [u8; 32] ) -> DispatchResult

Proposal removal

Tries to remove a specified proposal. Only the user who created the proposal can remove it.

Parameters:

proposal_id: the proposal identifier

pub fn set_bdk_url( origin: OriginFor<T>, new_url: BoundedVec<u8, ConstU32<32>> ) -> DispatchResult

BDK URL insertion

Changes the BDK-services endpoint, useful for pointing to the btc mainnet or testnet

Parameters:

new_url: The new endpoint to which all the bdk related requests will be sent.

Considerations

Ensure the new url is valid.

The url has a maximum length of 32 bytes

pub fn save_psbt( origin: OriginFor<T>, proposal_id: [u8; 32], signature_payload: BoundedVec<u8, T::PSBTMaxLen> ) -> DispatchResult

PSBT signature insertion

Stores the signature for a PSBT proposal

Parameters:

proposal_id: the proposal identifier

signature_payload: a blob of psbt bytes, resulting from a external wallet

Considerations

If successful, this process cannot be undone

A user can only sign a proposal once

pub fn finalize_psbt( origin: OriginFor<T>, proposal_id: [u8; 32], broadcast: bool ) -> DispatchResult

Finalize PSBT

Queries a proposal to be finalized generating a tx_id in the process, it can also be broadcasted if specified.

Parameters:

proposal_id: the proposal identifier

broadcast: A boolean flag

Considerations

If successful, this process cannot be undone

The proposal must have a valid PSBT

Any vault member can perform this extrinsic

pub fn broadcast_psbt( origin: OriginFor<T>, proposal_id: [u8; 32] ) -> DispatchResult

Broadcast PSBT

Queries a proposal to be broadcasted in case it wasn’t on the finalization step.

Parameters:

proposal_id: the vault identifier in which the proposal will be inserted

Considerations

If successful, this process cannot be undone

The proposal must be finalized already

Any vault member can perform this extrinsic

pub fn create_proof( origin: OriginFor<T>, vault_id: [u8; 32], message: Description<T>, psbt: PSBT<T> ) -> DispatchResult

Create Proof of Reserve

Stores a PoR for a defined vault.

Parameters:

vault_id: the vault identifier in which the proof will be inserted

message: the message to be taken into account to generate the PoR PSBT

psbt: the psbt generated from bdk

Considerations:

Any vault member can perform this extrinsic

A vault can only have a PoR at a time.

pub fn save_proof_psbt( origin: OriginFor<T>, vault_id: [u8; 32], psbt: PSBT<T> ) -> DispatchResult

Save Proof of reserve PSBT

Updates the PoR with a new PSBT

Parameters

vault_id: the vault identifier in which the proof is

psbt: the new psbt to insert, the signer will be linked to it.

Considerations:

Any vault member can perform this extrinsic

A vault signer can only sabe its PSBT once.

pub fn finalize_proof( origin: OriginFor<T>, vault_id: [u8; 32], psbt: PSBT<T> ) -> DispatchResult

Finalize proof of reserve

Updates the PoR with the final PSBT

Parameters

vault_id: the vault identifier in which the proof is

psbt: the new psbt to insert, the signer will be linked to it.

Considerations:

Any vault member can perform this extrinsic

A vault signer can only sabe its PSBT once.

pub fn kill_storage(origin: OriginFor<T>) -> DispatchResult

Kill almost all storage

Use with caution!

Can only be called by root and removes All vaults and proposals

pub fn ocw_insert_descriptors( origin: OriginFor<T>, payload: VaultsPayload<T::Public>, _signature: T::Signature ) -> DispatchResult

Extrinsic to insert a valid vault descriptor

Meant to be unsigned with signed payload and used by an offchain worker

pub fn ocw_insert_psbts( origin: OriginFor<T>, payload: ProposalsPayload<T::Public>, _signature: T::Signature ) -> DispatchResult

Extrinsic to insert a valid proposal PSBT

Meant to be unsigned with signed payload and used by an offchain worker

pub fn ocw_finalize_psbts( origin: OriginFor<T>, payload: ProposalsPayload<T::Public>, _signature: T::Signature ) -> DispatchResult

Extrinsic to insert a valid proposal TX_ID

Meant to be unsigned with signed payload and used by an offchain worker

impl<T: Config> Pallet<T>

pub fn xpubs<KArg>(k: KArg) -> Option<BoundedVec<u8, T::XPubLen>>where KArg: EncodeLike<[u8; 32]>,

Stores hash-xpub pairs

impl<T: Config> Pallet<T>

pub fn xpubs_by_owner<KArg>(k: KArg) -> Option<[u8; 32]>where KArg: EncodeLike<T::AccountId>,

Keeps track of what accounts own what xpub.

impl<T: Config> Pallet<T>

pub fn proposals<KArg>(k: KArg) -> Option<Proposal<T>>where KArg: EncodeLike<[u8; 32]>,

impl<T: Config> Pallet<T>

pub fn proposals_by_vault<KArg>( k: KArg ) -> BoundedVec<[u8; 32], T::MaxProposalsPerVault>where KArg: EncodeLike<[u8; 32]>,

impl<T: Config> Pallet<T>

pub fn vaults<KArg>(k: KArg) -> Option<Vault<T>>where KArg: EncodeLike<[u8; 32]>,

Stores vaults

impl<T: Config> Pallet<T>

pub fn vaults_by_signer<KArg>( k: KArg ) -> BoundedVec<[u8; 32], T::MaxVaultsPerUser>where KArg: EncodeLike<T::AccountId>,

Keeps track of what accounts own what xpub.

impl<T: Config> Pallet<T>

pub fn proof_of_reserve<KArg>(k: KArg) -> Option<ProofOfReserve<T>>where KArg: EncodeLike<[u8; 32]>,

Stores vaults proof-of-reserve