Trait pallet_bitcoin_vaults::pallet::Config

pub trait Config: Config + CreateSignedTransaction<Call<Self>> {
    type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type ChangeBDKOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type XPubLen: Get<u32>;
    type PSBTMaxLen: Get<u32>;
    type MaxVaultsPerUser: Get<u32>;
    type MaxCosignersPerVault: Get<u32>;
    type VaultDescriptionMaxLen: Get<u32>;
    type OutputDescriptorMaxLen: Get<u32>;
    type MaxProposalsPerVault: Get<u32>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type AuthorityId: AppCrypto<Self::Public, Self::Signature>

The identifier type for an offchain worker.

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type ChangeBDKOrigin: EnsureOrigin<Self::RuntimeOrigin>

type XPubLen: Get<u32>

type PSBTMaxLen: Get<u32>

type MaxVaultsPerUser: Get<u32>

It counts both owned vaults and vaults where the user is a cosigner

type MaxCosignersPerVault: Get<u32>

type VaultDescriptionMaxLen: Get<u32>

type OutputDescriptorMaxLen: Get<u32>

type MaxProposalsPerVault: Get<u32>